Move lottie_convert.py conversion code to its own file
